+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed featured post duplication by filtering non-numeric values from exclusion arrays and adding explicit `post_status` filter
 - RSS feed title and post author incorrect values
+- Strip tags from page title html
 
 ## [4.3.0] - 2026-01-07